# cet252/APIDOC/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Music Library API",
    "version": "1.0.0",
    "description": "REST API for managing tracks in a music library"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check endpoint",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "API is available"
          }
        }
      }
    },
    "/api/tracks": {
      "get": {
        "summary": "Retrieve all tracks",
        "tags": [
          "Tracks"
        ],
        "responses": {
          "200": {
            "description": "A list of tracks"
          }
        }
      },
      "post": {
        "summary": "Create a new track",
        "tags": [
          "Tracks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Track created"
          },
          "400": {
            "description": "Invalid payload"
          }
        }
      }
    },
    "/api/tracks/{id}": {
      "get": {
        "summary": "Retrieve a track by id",
        "tags": [
          "Tracks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track found"
          },
          "404": {
            "description": "Track not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing track",
        "tags": [
          "Tracks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Track updated"
          },
          "404": {
            "description": "Track not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a track by id",
        "tags": [
          "Tracks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track deleted"
          },
          "404": {
            "description": "Track not found"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}